Updates to wayland protocol test framework
1) Move wayland test functions to helper
2) Add ability to clone test client
