+++ test/backend/wayland/test_session_lock.py
@@ -1,10 +1,14 @@
+from functools import partial
 from types import MethodType
 
 import pytest
 
 from libqtile.backend.wayland._ffi import lib
 from libqtile.command.base import CommandError
+from test.backend.wayland.helpers import count_node_types, find_layer, find_node_at_position
 from test.conftest import dualmonitor
+
+find_layer_lock = partial(find_layer, layer_name="LAYER_LOCK")
 
 
 class LockState:
@@ -13,76 +17,6 @@ class LockState:
     LOCKED = lib.QW_SESSION_LOCK_LOCKED
     UNLOCKED = lib.QW_SESSION_LOCK_UNLOCKED
     CRASHED = lib.QW_SESSION_LOCK_CRASHED
-
-
-# Functions for finding/navigating LAYER_LOCK
-
-
-def find_layer_lock(stacking_info):
-    """Recursively search the nested stacking tree for the LAYER_LOCK node."""
-    if not isinstance(stacking_info, dict):
-        return None
-
-    # Check current node
-    if stacking_info.get("name") == "LAYER_LOCK":
-        return stacking_info
-
-    # Recurse through children
-    for child in stacking_info.get("children", []):
-        result = find_layer_lock(child)
-        if result is not None:
-            return result
-
-    return None
-
-
-def find_node_at_position(node, node_type, x, y, parent_x=0, parent_y=0):
-    """
-    Looks for a node at given position but, as tree nodes are relative,
-    we need to track absolute position of child node.
-    """
-    if not isinstance(node, dict):
-        return False
-
-    abs_x = parent_x + node["x"]
-    abs_y = parent_y + node["y"]
-
-    if node["type"] == node_type:
-        return abs_x == x and abs_y == y
-
-    for child in node.get("children", []):
-        result = find_node_at_position(
-            child,
-            node_type,
-            x,
-            y,
-            abs_x,
-            abs_y,
-        )
-        if result:
-            return True
-
-    return False
-
-
-def count_node_types(node, counts=None):
-    """Returns dict of the number of each node type in a given node."""
-    if counts is None:
-        counts = {}
-
-    if not isinstance(node, dict):
-        return counts
-
-    # Count current node type
-    node_type = node.get("type")
-    if node_type:
-        counts[node_type] = counts.get(node_type, 0) + 1
-
-    # Recurse through children
-    for child in node.get("children", []):
-        count_node_types(child, counts)
-
-    return counts
 
 
 @pytest.fixture
